+++ config/config.go
@@ -1,4 +1,4 @@
-package main
+package exporter
 
 import (
 	"flag"
