+++ prusa_link/api.go
@@ -1,4 +1,4 @@
-package main
+package exporter
 
 import (
 	"errors"
