+++ prusa_link/sl_structs.go
@@ -1,4 +1,4 @@
-package main
+package prusalink
 
 // SLJob is a struct to hold the response from the job API endpoint
 type SLJob struct {
