+++ prusa_link/xBuddy.go
@@ -1,4 +1,4 @@
-package main
+package prusalink
 
 import (
 	"strings"
