+++ prusa_link/xBuddy_scrape.go
@@ -1,4 +1,4 @@
-package main
+package prusalink
 
 import (
 	"encoding/json"
