+++ syslog/syslog.go
@@ -1,4 +1,4 @@
-package main
+package syslog
 
 import (
 	"bytes"
